+++ app/ui/startup/browser.js
@@ -1,5 +1,7 @@
 //(function () {'use strict';
 
+const os = require('os');
+
 const electron = require('electron');
 const ipcRenderer = electron.ipcRenderer;
 const handlebars = require('handlebars');
@@ -14,6 +16,8 @@ const app = electron.remote.app;
 
 const logger = require('../../lib/logger.js');
 const loggerFactory = require('../../lib/logger-factory.js');
+const globalConfig = require('../../lib/global-config.js');
+
 var standardLogger = new loggerFactory(clientConfig.getAll());
 logger.setLogger(standardLogger);
 
@@ -132,7 +136,15 @@ function verifyServer() {
 function pingApiServer(blnUrl, callback) {
   var apiPingUrl = blnUrl + '/api/v2';
 
-  request.get(apiPingUrl, {timeout: 2000}, (err, result) => {
+  var reqOptions = {
+    timeout: 2000,
+    headers: {
+      'X-Client': ['Balloon-Desktop-App', globalConfig.get('version'), os.hostname()].join('|'),
+      'User-Agent': ['Balloon-Desktop-App', globalConfig.get('version'), os.hostname(), os.platform(), os.release()].join('|'),
+    }
+  };
+
+  request.get(apiPingUrl, reqOptions, (err, result) => {
     try {
       var body = JSON.parse(result.body);
       callback(!(err || body.name !== 'balloon'));
@@ -174,13 +186,26 @@ function advanced() {
 }
 
 function auth() {
-  if(env.auth && env.auth.basic === false) {
-    $('#startup-auth-basic').hide();
+  if(env.auth && env.auth.credentials === null) {
+    $('#startup-auth-credentials').hide();
   }
 
   var $loader = $('.window-loader');
   var $container = $('#startup-auth-oidc');
   $container.find('> img').remove();
+
+  var $username = $('#startup-view-auth').find('input[name=username]');
+  var $password = $('#startup-view-auth').find('input[name=password]');
+  var $mfaCode = $('#startup-view-auth').find('input[name=mfaCode]');
+  var $description = $('#startup-view-auth').find('p:first-child');
+
+  $username.show();
+  $password.show();
+  $mfaCode.hide();
+
+  $username.val('');
+  $password.val('');
+  $mfaCode.val('');
 
   if(env.auth && env.auth.oidc) {
     var i=0;
@@ -195,30 +220,38 @@ function auth() {
     ipcRenderer.send('auth-oidc-signin', $(this).attr('alt'));
   });
 
+  ipcRenderer.removeAllListeners('startup-auth-mfa-required');
+  ipcRenderer.on('startup-auth-mfa-required', function(event) {
+    $loader.hide();
+
+    if(env.auth && ['basic', null].includes(env.auth.credentials)) {
+      $('#startup-auth-error').find('> div').hide()
+      $('#startup-auth-error-mfa-not-token').show();
+    } else {
+      $username.hide();
+      $password.hide();
+      $('#startup-auth-error').find('> div').hide();
+      $mfaCode.show().focus();
+      $description.html(i18n.__('startup.auth.mfaDescription'));
+    }
+  });
+
+  ipcRenderer.removeAllListeners('startup-auth-error');
   ipcRenderer.on('startup-auth-error', function (event, type) {
     $loader.hide();
     $('#startup-auth-error').find('> div').hide()
     $('#startup-auth-error-'+type).show();
   });
 
-  function basicAuth() {
+  $('#startup-auth-credentials').off('submit').on('submit', function(event) {
+    event.preventDefault();
+
     $loader.show();
-    var username = $('#startup-view-auth').find('input[name=username]').val();
-    var password = $('#startup-view-auth').find('input[name=password]').val();
-    ipcRenderer.send('startup-basic-auth', username, password);
-  }
-
-  $(document).bind('keypress', function(e){
-    if($(e.target).attr('id') === 'startup-auth-continue') {
-      return;
-    }
-
-    if(e.which === 13 && $('#startup-view-auth').is(':visible')) {
-      basicAuth(e);
-    }
+    var username = $username.val();
+    var password = $password.val();
+    var code = $mfaCode.val();
+    ipcRenderer.send('startup-credentials-signin', username, password, code);
   });
-
-  $('#startup-auth-continue').bind('click', basicAuth);
 }
 
 function switchView(view) {